+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 # 3.6.0 (XX.XX.XXXX)
 
 ### Features
-* **tc-checkbox-group** added new input `collapsible` which - if set to true - collapses the component to its header initially.
-* **terra-multi-check-box** added new input `collapsible` which - if set to true - collapses the component to its header initially if set to true.
+* **tc-checkbox-group** is now collapsible. Added new input `collapsed` which - if set to true - collapses the component to its header initially.
+* **terra-multi-check-box** is now collapsible. Added new input `collapsed` which - if set to true - collapses the component to its header initially.
 
 <a name="3.5.1"></a>
 # 3.5.1 (20.02.2019)